update content, split sections into separate components for better readability
